First cut/strawman at adding cities
I added a fav.css, favortieCities.html and the javascript code (fav.css). to get this part started.  I also created the firebase database.  Right now it just adds cities, lists currently stored favorite cities and allows yoou to delete cities.  So, its just a start but wanted to get that out there so it could be looked at and connected if needed.