Updated favorites maintenance page to comply
I updated the html page for adding favorites to comply with Mike's style in the dsahboard so they look amnd feel the same.  I am have some issues during *update* of a table row that I need to get help on.   The row wants to go outside the cotainer for the card-body or it wants to force the user to scroll left and right to get at the info.  But it works.  I also moves a couple of the unused files into an unused directory so it does not pollute our filesystem with artifcats we are not using.