Several navigation/data passing updates
I updated the pages to pass data between the index.html page and the edit controller and page.
1.  Every time a favorite is selected, I store in in local storage to be used later
2. when page loads I get last fav used from local storage and use that as the selected location for calling weather and places services
3. when a user selects a favorite on the EDIT page, I store that favorite and pass it to the dashboard page.

So switching bewteen pages works. But I still would like to add a *close* button to the card so that it will just jump back and leave the storage alone.  That is next commit